+++ lib/oli_web/controllers/project_controller.ex
@@ -9,6 +9,7 @@ defmodule OliWeb.ProjectController do
   alias Oli.Qa
   alias Oli.Analytics.Datashop
   alias OliWeb.Common.Breadcrumb
+  alias Oli.Authoring.Clone
   alias Oli.Activities
 
   def overview(conn, project_params) do
@@ -173,4 +174,17 @@ defmodule OliWeb.ProjectController do
     )
   end
 
+  def clone_project(conn, _project_params) do
+    case Clone.clone_project(conn.assigns.project.slug, conn.assigns.current_author) do
+      {:ok, project} ->
+        conn
+        |> put_flash(:info, "Project duplicated. You've been redirected to your new project.")
+        |> redirect(to: Routes.project_path(conn, :overview, project))
+
+      {:error, message} ->
+        conn
+        |> put_flash(:error, "Project could not be copied: " <> message)
+        |> render("overview.html")
+    end
+  end
 end